Expose Cloned HTML to Hooks Fixes #59
