+++ src/lib/Item.ts
@@ -1,4 +1,5 @@
 import { Clone, Template } from "src/clones";
+import Lifecycle from "./Lifecycle";
 
 function* genId(): Generator<number, any> {
     let id = 0;
@@ -11,10 +12,13 @@ export default class Item {
     private static genId = genId();
 
     private _id: number;
+    private _lifecycle: Lifecycle;
+
     protected clone: Clone;
 
-    constructor(template: Template) {
+    constructor(template: Template, lifecycle: Lifecycle = new Lifecycle()) {
         this.clone = new Clone(template);
+        this._lifecycle = lifecycle;
 
         this._id = Item.genId.next().value;
         this.clone.id = this._id;
@@ -41,5 +45,15 @@ export default class Item {
 
     remove() {
         this.clone.remove();
+    }
+
+    onCreated() {
+        this.clone.onCreated(this._lifecycle.onCreated);
+    }
+    onDestroyed() {
+        this.clone.onDestroyed(this._lifecycle.onDestroyed);
+    }
+    each() {
+        this.clone.each(this._lifecycle.each);
     }
 }
